# backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/summary/{uuid}:
    post:
      tags:
      - chat-summary-controller
      operationId: startSummary
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryRequestDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskAckResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/summary/restart/{uuid}:
    post:
      tags:
      - chat-summary-controller
      operationId: restartSummary
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryRequestDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskAckResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/summary/pause/{uuid}:
    post:
      tags:
      - chat-summary-controller
      operationId: pauseSummary
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskAckResponse"
...
  /api/summary/status-pool/{uuid}:
    get:
      tags:
      - chat-summary-controller
      operationId: getStatusAndProgress
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSummaryProgressResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/summary/preview/{uuid}:
    get:
      tags:
      - chat-summary-controller
      operationId: getChatPreview
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseChatPreviewResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      description: Unified error body. The code mirrors the HTTP status.
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    ApiResponseChatPreviewResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/ChatPreviewResponse"
...
    ApiResponseSummaryProgressResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/SummaryProgressResponse"
    ApiResponseTaskAckResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/TaskAckResponse"
...
    SummaryRequestDTO:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
...
    ChatPreviewResponse:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties:
            type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ChatPreviewRow"
    ChatPreviewRow:
      type: object
      properties:
        lineId:
          type: string
        timestamp:
          type: string
        sender:
          type: string
        content:
          type: string
    SummaryProgressResponse:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - INITIAL_STATE
          - RUNNING
          - PAUSED
          - SUCCESS
          - FAILED
        progress:
          type: number
          format: double
        result:
          type: string
        errorMessage:
          type: string
    TaskAckResponse:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
          nullable: true
          enum:
          - INITIAL_STATE
          - RUNNING
          - PAUSED
          - SUCCESS
          - FAILED
        message:
          type: string
...